expense-bot-server/bot/requirements.txt
Flask
google-generativeai
flask_cors
psycopg2-binary
apscheduler
Flask-JWT-Extended
bcrypt
python-dotenv
sqlalchemy